test: 로또 당첨 계산 WinningNumbers 클래스로 이동으로 인한 TC 수정
