+++ src/test/java/lotto/domain/winning/WinningNumbersTest.java
@@ -1,5 +1,7 @@
 package lotto.domain.winning;
 
+import lotto.domain.generator.FixedNumberGenerator;
+import lotto.domain.generator.NumberGenerator;
 import lotto.domain.number.LottoNumber;
 import lotto.domain.number.LottoNumbers;
 import org.junit.jupiter.api.DisplayName;
@@ -32,5 +34,19 @@ public class WinningNumbersTest {
     @Test
     void equals() {
         assertThat(winningNumbers.equals(new WinningNumbers(winningNumberString))).isTrue();
+    }
+
+    @DisplayName("LottoNumbers 비교하여 존재하는 갯수를 리턴할 수 있다.")
+    @Test
+    void matchCount() {
+        String winningNumberString = "1,2,3,4,5,6";
+        WinningNumbers winningNumbers = new WinningNumbers(winningNumberString);
+
+        NumberGenerator<LottoNumber> fixedNumberGenerator = new FixedNumberGenerator();
+        LottoNumbers lottoNumbers = LottoNumbers.newLottoNumbersWithNumberGenerator(fixedNumberGenerator);
+
+        int matchCount = winningNumbers.matchCount(lottoNumbers);
+
+        assertThat(matchCount).isEqualTo(fixedNumberGenerator.getNumbers().size());
     }
 }
